Refactor: Extract data loading and filtering logic into utility module
- Create utils/collegeDataLoader.js with data loading, filtering, and sorting functions
- Refactor pages/api/exam-result.js to use the new utility module
- Improve code maintainability by separating concerns
- Maintain backward compatibility - API responses unchanged
- Fixes issue #159
