+++ pages/api/exam-result.js
@@ -1,6 +1,11 @@
-import fs from "fs/promises";
 import examConfigs from "../../examConfig";
 import rateLimit from "express-rate-limit";
+import {
+  loadCollegeData,
+  createRankFilter,
+  applyFilters,
+  sortColleges,
+} from "../../utils/collegeDataLoader";
 
 // Helper function to get client IP address
 const getIp = (req) => {
@@ -106,158 +111,13 @@ export default async function handler(req, res) {
 
   try {
     const dataPath = config.getDataPath(req.query.category);
-    const data = await fs.readFile(dataPath, "utf8");
-    const fullData = JSON.parse(data);
+    const fullData = await loadCollegeData(dataPath);
 
-    // Get filters based on the exam config and query parameters
     const filters = config.getFilters(req.query);
+    const rankFilter = createRankFilter(exam, req.query);
 
-    // Helper function to parse rank (handles 'P' suffix)
-    const parseRank = (rankStr) => {
-      if (!rankStr) return null;
-      const numStr = rankStr.toString().replace(/[^0-9]/g, "");
-      return numStr ? parseInt(numStr, 10) : null;
-    };
-
-    const hasPSuffix = (rankStr) => {
-      if (!rankStr) return false;
-      return rankStr.toString().trim().toUpperCase().endsWith("P");
-    };
-
-    const rankFilter = (item) => {
-      if (exam === "GUJCET") {
-        // For GUJCET, filter based on closing_marks
-        const cutoffMarks = parseFloat(item.closing_marks) || 0;
-        const userMarks = parseFloat(rank) || 0;
-        return userMarks >= cutoffMarks * 0.9; // Show if user marks are >= 90% of cutoff
-      } else if (exam === "NEET") {
-        // For NEET, filter based on closing rank with 0.9 coefficient
-        const closingRank = parseFloat(item["Closing Rank"]) || 0;
-        const userRank = parseFloat(rank) || 0;
-        return closingRank >= 0.9 * userRank; // Show colleges where closing rank is >= 90% of user's rank
-      } else if (exam === "TNEA") {
-        return parseFloat(item["Cutoff Marks"]) <= parseFloat(rank);
-      }
-
-      const itemRankStr = item["Closing Rank"]?.toString().trim() || "";
-      const itemRank = parseRank(itemRankStr);
-      const itemHasPSuffix = hasPSuffix(itemRankStr);
-
-      if (exam === "JoSAA") {
-        if (item["Exam"] === "JEE Advanced") {
-          if (req.query.qualifiedJeeAdv !== "Yes" || !req.query.advRank)
-            return false;
-
-          const userRankStr = req.query.advRank?.toString().trim() || "";
-          const userRank = parseRank(userRankStr);
-          const userHasPSuffix = hasPSuffix(userRankStr);
-
-          // If one has 'P' suffix and the other doesn't, they don't match
-          if (itemHasPSuffix !== userHasPSuffix) return false;
-
-          return userRank && itemRank >= 0.9 * userRank;
-        } else {
-          if (!req.query.mainRank) return false;
-
-          const userRankStr = req.query.mainRank?.toString().trim() || "";
-          const userRank = parseRank(userRankStr);
-          const userHasPSuffix = hasPSuffix(userRankStr);
-
-          // If one has 'P' suffix and the other doesn't, they don't match
-          if (itemHasPSuffix !== userHasPSuffix) return false;
-
-          return userRank && itemRank >= 0.9 * userRank;
-        }
-      } else if (exam === "JEE Advanced") {
-        if (item["Exam"] !== "JEE Advanced") return false;
-        if (!req.query.advRank) return false;
-
-        const userRankStr = req.query.advRank?.toString().trim() || "";
-        const userRank = parseRank(userRankStr);
-        const userHasPSuffix = hasPSuffix(userRankStr);
-
-        // If one has 'P' suffix and the other doesn't, they don't match
-        if (itemHasPSuffix !== userHasPSuffix) return false;
-
-        return userRank && itemRank >= 0.9 * userRank;
-      } else if (exam === "JEE Main") {
-        if (item["Exam"] === "JEE Advanced") return false;
-        if (!req.query.mainRank) return false;
-
-        const userRankStr = req.query.mainRank?.toString().trim() || "";
-        const userRank = parseRank(userRankStr);
-        const userHasPSuffix = hasPSuffix(userRankStr);
-
-        // If one has 'P' suffix and the other doesn't, they don't match
-        if (itemHasPSuffix !== userHasPSuffix) return false;
-
-        return userRank && itemRank >= 0.9 * userRank;
-      } else {
-        return true;
-      }
-    };
-
-    let filteredData = fullData;
-
-    // Apply filters if they exist
-    if (Array.isArray(filters)) {
-      filteredData = filteredData.filter((item) => {
-        return filters.every((filterFn) => filterFn(item));
-      });
-    }
-
-    // Apply rank filter if it exists
-    if (rankFilter) {
-      filteredData = filteredData.filter(rankFilter);
-    }
-
-    // Apply sorting based on exam type
-    if (config.getSort) {
-      const sortConfig = config.getSort();
-      if (sortConfig && sortConfig.length > 0) {
-        const [sortKey, sortOrder] = sortConfig[0];
-        filteredData.sort((a, b) => {
-          let valA = a[sortKey];
-          let valB = b[sortKey];
-
-          // Convert to numbers if possible for proper numeric comparison
-          if (!isNaN(parseFloat(valA)) && !isNaN(parseFloat(valB))) {
-            valA = parseFloat(valA);
-            valB = parseFloat(valB);
-          }
-
-          if (sortOrder === "DESC") {
-            return valB - valA;
-          }
-          return valA - valB;
-        });
-      }
-    } else if (exam === "TGEAPCET") {
-      // For TGEAPCET, sort by closing_rank in ascending order
-      filteredData.sort((a, b) => {
-        const rankA = parseInt(a.closing_rank, 10) || 0;
-        const rankB = parseInt(b.closing_rank, 10) || 0;
-        return rankA - rankB; // Ascending order (lower ranks first)
-      });
-    } else if (exam === "TNEA") {
-      // For TNEA, sort by cutoff marks in descending order
-      filteredData.sort((collegeA, collegeB) => {
-        const collegeAMarks = parseFloat(collegeA["Cutoff Marks"]) || 0;
-        const collegeBMarks = parseFloat(collegeB["Cutoff Marks"]) || 0;
-        return collegeBMarks - collegeAMarks; // Descending order (higher cutoff first)
-      });
-    } else if (
-      exam === "JEE Main" ||
-      exam === "JEE Advanced" ||
-      exam === "JoSAA"
-    ) {
-      // For JEE Main, JEE Advanced, and JoSAA, sort by Closing Rank ascending
-      filteredData.sort((collegeA, collegeB) => {
-        const rankA = parseFloat(collegeA["Closing Rank"]) || 0;
-        const rankB = parseFloat(collegeB["Closing Rank"]) || 0;
-        return rankA - rankB;
-      });
-    }
+    let filteredData = applyFilters(fullData, filters, rankFilter);
+    filteredData = sortColleges(filteredData, exam, config);
 
     return res.status(200).json(filteredData);
   } catch (error) {
